Refactor card list into its own component
