+++ components/IssuesCardListErrorScreen.vue
@@ -12,7 +12,7 @@
 import { Vue, Component } from 'nuxt-property-decorator'
 
 @Component
-export default class IssuesErrorScreen extends Vue {}
+export default class IssuesCardListErrorScreen extends Vue {}
 </script>
 
 <style scoped lang="scss">
